requirements.txt
amqp==5.2.0
anyio==4.3.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
billiard==4.2.0
black==24.2.0
celery==5.3.6
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.0.2
django-debug-toolbar==4.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
flake8==7.0.0
h11==0.14.0
httpcore==1.0.4
httpx==0.26.0
idna==3.6
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kombu==5.3.5
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
pycodestyle==2.11.1
pyflakes==3.2.0
PyJWT==2.8.0
pyTelegramBotAPI==4.16.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==20.8
pytz==2024.1
PyYAML==6.0.1
redis==5.0.2
referencing==0.33.0
requests==2.31.0
rpds-py==0.18.0
six==1.16.0
slugify==0.0.1
sniffio==1.3.1
sqlparse==0.4.4
stripe==8.5.0
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
uuid==1.30
vine==5.1.0
wcwidth==0.2.13
